Day 17 part 2.
I couldn't speed up the Elixir solution for part 2 so re-implemented it
in Python. The Python solution is also pretty slow, but fast enough.

The only significant difference is that the Python solution uses a
priority queue rather than a list, which makes finding the next node to
visit faster.